requirements.txt
streamlit>=1.35.0
google-generativeai>=0.3.0
crawl4ai>=0.3.0
playwright>=1.40.0
plotly>=5.18.0
pandas>=2.1.0